==> bin/scripts/0_setup.R <==
# These are the packages required for MicroCyte R scripts:

install.packages("ggplot2")
install.packages("ggpubr")
install.packages("MASS")
install.packages("viridis")
install.packages("stringr")
devtools::install_github("vqv/ggbiplot")
install.packages("imager")
install.packages("magick")
